+++ CHANGELOG.rst
@@ -1,5 +1,7 @@
 1.3-edw2 (unreleased)
 -------------------
+* added verbose paramenter to return meaningful errors instead of just
+  False [dumitval]
 
 1.3-edw1 (2016-08-10)
 -------------------